Update target Debian distribution to Bullseye (11).
Although Debian Bullseye is not yet released, the initial freeze phase
has begun and the official release will likely be sometime in the first half of
2021.

Bullseye is currently available as the Debian [testing][1] distribution
which means that it's already possible to install and use directly and
docker container images (as would be used on the buildfarm if CI or
binary packages were provided) are already available.
The testing distribution is not managed by the Debian Security team
which only means that security updates are not made urgently but this
will change with the official release of Bullseye.

Aside from the general updates from one Debian release to the next one
major motivation to target Bullseye over Buster is the absense of the
python3-importlib-metadata package in Buster.

[1]: https://www.debian.org/releases/testing/