Remove acceleration_firmware common firmware front-end
Refer to https://github.com/ros-acceleration/acceleration_firmware/issues/1.

Shortly, the rationale of having was to provide
a mechanism to splitting vendors across firmware
packages while keeping a common front-end. After
looking at it critically, there's currently no
strong argument to keep this abstract layer at this
level since this same behavior can be obtained through
colcon-hardware-acceleration extensions so for now,
we've decided to remove this layer from the architecture.

In the future, if needed, acceleration_firmware will be
re-added and should include vendor-common scripts to
deploy and manage firmware

Signed-off-by: Víctor Mayoral Vilches <v.mayoralv@gmail.com>