Use the platform package name format for independent variants.
The independent variants are not implemented as ROS packages with
package.xml name restrictions and their transformation into platform
package names.

Because of that I think it makes the most sense to refer to them
exclusively by their platform package names since those will likely be
the names of the rosdep keys used for them.